+++ pom.xml
@@ -10,8 +10,8 @@
 
     <groupId>io.github.sinri</groupId>
     <artifactId>Keel</artifactId>
-    <version>3.0.1-SNAPSHOT</version>
-    <!--    <version>3.0.0</version>-->
+    <version>3.0.2-SNAPSHOT</version>
+    <!--        <version>3.0.1</version>-->
 
     <name>Keel</name>
     <description>
