+++ src/main/java/io/github/sinri/keel/mysql/KeelMySQLConfiguration.java
@@ -7,6 +7,7 @@ import io.vertx.sqlclient.PoolOptions;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.Objects;
+import java.util.concurrent.TimeUnit;
 
 /**
  * KeelMySQLConfigure for connections and pool.
@@ -77,6 +78,12 @@ public class KeelMySQLConfiguration extends KeelConfiguration {
         if (schema != null) {
             mySQLConnectOptions.setDatabase(schema);
         }
+
+        Integer connectionTimeout = getConnectionTimeout();
+        if (connectionTimeout != null) {
+            mySQLConnectOptions.setConnectTimeout(connectionTimeout);
+        }
+
         return mySQLConnectOptions;
     }
 
@@ -87,9 +94,10 @@ public class KeelMySQLConfiguration extends KeelConfiguration {
         if (poolMaxSize != null) {
             poolOptions.setMaxSize(poolMaxSize);
         }
-        Integer connectionTimeout = getConnectionTimeout();
-        if (connectionTimeout != null) {
-            poolOptions.setConnectionTimeout(connectionTimeout);
+        Integer poolConnectionTimeout = getPoolConnectionTimeout();
+        if (poolConnectionTimeout != null) {
+            poolOptions.setConnectionTimeout(poolConnectionTimeout);
+            poolOptions.setConnectionTimeoutUnit(TimeUnit.SECONDS);
         }
         return poolOptions;
     }
@@ -134,7 +142,24 @@ public class KeelMySQLConfiguration extends KeelConfiguration {
         return dataSourceName;
     }
 
-    public Integer getConnectionTimeout() {
+    /**
+     * The default value of connect timeout = 60000 ms
+     *
+     * @return connectTimeout - connect timeout, in ms
+     * @since 3.0.1 let it be its original setting!
+     */
+    private Integer getConnectionTimeout() {
         return readAsInteger("connectionTimeout");
+    }
+
+    /**
+     * Set the amount of time a client will wait for a connection from the pool.
+     * If the time is exceeded without a connection available, an exception is provided.
+     * TimeUnit would be set by `setConnectionTimeoutUnit`
+     *
+     * @see <a href="https://vertx.io/docs/apidocs/io/vertx/sqlclient/PoolOptions.html#setConnectionTimeout-int-">...</a>
+     */
+    public Integer getPoolConnectionTimeout() {
+        return readAsInteger("poolConnectionTimeout");
     }
 }
